+++ execution_engine/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.  The format
 
 ### Added
 * Added genesis validation step to ensure there are more genesis validators than validator slots.
+* Added a `max_associated_keys` configuration option for a hard limit of associated keys under accounts.